+++ Command/QueryCommand.php
@@ -9,7 +9,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Helper\TableHelper;
 
-class MassiveSearchQueryCommand extends ContainerAwareCommand
+class QueryCommand extends ContainerAwareCommand
 {
     public function configure()
     {