+++ Tests/Functional/Command/QueryCommandTest.php
@@ -5,14 +5,14 @@ namespace Massive\Bundle\SearchBundle\Tests\Functional;
 use Massive\Bundle\SearchBundle\Tests\Resources\TestBundle\Entity\Product;
 use Massive\Bundle\SearchBundle\Tests\Resources\TestBundle\EventSubscriber\TestSubscriber;
 use Symfony\Component\Console\Tester\CommandTester;
-use Massive\Bundle\SearchBundle\Command\MassiveSearchQueryCommand;
+use Massive\Bundle\SearchBundle\Command\QueryCommand;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 
-class MassiveSearchQueryTest extends BaseTestCase
+class QueryCommandTest extends BaseTestCase
 {
     public function setUp()
     {
-        $command = new MassiveSearchQueryCommand();
+        $command = new QueryCommand();
         $application = new Application($this->getContainer()->get('kernel'));
         $command->setApplication($application);
         $this->tester = new CommandTester($command);