+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 dev-master
 ----------
 
+- Added `massive:search:status` command to show information about the current implementation
 - Add document ReflectionClass to HitEvent
 - Added `massive:search:query` command to execute queries from the CLI
 - XML driver handles meta fields Url, Title and Description